@@ -3,8 +3,8 @@ apply plugin: 'java'
 apply plugin: 'maven-publish'
 
 compileJava {
-  sourceCompatibility '11'
-  targetCompatibility '11'
+  sourceCompatibility '17'
+  targetCompatibility '17'
 }
 
 group = 'net.fabricmc'
@@ -24,7 +24,7 @@ repositories { mavenCentral() }
 dependencies {
   implementation 'org.jetbrains:annotations:20.1.0'
   testImplementation 'junit:junit:4.12'
-  testImplementation 'org.assertj:assertj-core:3.12.2'
+  testImplementation 'org.assertj:assertj-core:3.23.1'
 }
 
 jar {
