@@ -1,6 +1,7 @@
 // Copyright 2000-2021 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 package org.jetbrains.java.decompiler.modules.decompiler;
 
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.java.decompiler.code.CodeConstants;
 import org.jetbrains.java.decompiler.code.Instruction;
 import org.jetbrains.java.decompiler.code.InstructionSequence;
@@ -124,15 +125,11 @@ public class FinallyProcessor {
     BasicBlock firstBasicBlock = firstBlockStatement.getBlock();
     Instruction instrFirst = firstBasicBlock.getInstruction(0);
 
-    int firstcode = 0;
-
-    switch (instrFirst.opcode) {
-      case CodeConstants.opc_pop:
-        firstcode = 1;
-        break;
-      case CodeConstants.opc_astore:
-        firstcode = 2;
-    }
+    int firstCode = switch (instrFirst.opcode) {
+      case CodeConstants.opc_pop -> 1;
+      case CodeConstants.opc_astore -> 2;
+      default -> 0;
+    };
 
     ExprProcessor proc = new ExprProcessor(methodDescriptor, varProcessor);
     proc.processStatement(root, cl);
@@ -140,12 +137,12 @@ public class FinallyProcessor {
     SSAConstructorSparseEx ssa = new SSAConstructorSparseEx();
     ssa.splitVariables(root, mt);
 
-    List<Exprent> lstExprents = firstBlockStatement.getExprents();
+    List<Exprent> expressions = firstBlockStatement.getExprents();
 
-    VarVersionPair varpaar = new VarVersionPair((VarExprent)((AssignmentExprent)lstExprents.get(firstcode == 2 ? 1 : 0)).getLeft());
+    VarVersionPair pair = new VarVersionPair((VarExprent)((AssignmentExprent)expressions.get(firstCode == 2 ? 1 : 0)).getLeft());
 
-    FlattenStatementsHelper flatthelper = new FlattenStatementsHelper();
-    DirectGraph dgraph = flatthelper.buildDirectGraph(root);
+    FlattenStatementsHelper flattenHelper = new FlattenStatementsHelper();
+    DirectGraph dgraph = flattenHelper.buildDirectGraph(root);
 
     LinkedList<DirectNode> stack = new LinkedList<>();
     stack.add(dgraph.first);
@@ -170,20 +167,19 @@ public class FinallyProcessor {
 
       boolean isTrueExit = true;
 
-      if (firstcode != 1) {
-
+      if (firstCode != 1) {
         isTrueExit = false;
 
         for (int i = 0; i < node.exprents.size(); i++) {
           Exprent exprent = node.exprents.get(i);
 
-          if (firstcode == 0) {
+          if (firstCode == 0) {
             List<Exprent> lst = exprent.getAllExprents();
             lst.add(exprent);
 
             boolean found = false;
             for (Exprent expr : lst) {
-              if (expr.type == Exprent.EXPRENT_VAR && new VarVersionPair((VarExprent)expr).equals(varpaar)) {
+              if (expr.type == Exprent.EXPRENT_VAR && new VarVersionPair((VarExprent)expr).equals(pair)) {
                 found = true;
                 break;
               }
@@ -192,8 +188,8 @@ public class FinallyProcessor {
             if (found) {
               found = false;
               if (exprent.type == Exprent.EXPRENT_EXIT) {
-                ExitExprent exexpr = (ExitExprent)exprent;
-                if (exexpr.getExitType() == ExitExprent.EXIT_THROW && exexpr.getValue().type == Exprent.EXPRENT_VAR) {
+                ExitExprent exit = (ExitExprent)exprent;
+                if (exit.getExitType() == ExitExprent.EXIT_THROW && exit.getValue().type == Exprent.EXPRENT_VAR) {
                   found = true;
                 }
               }
@@ -206,13 +202,12 @@ public class FinallyProcessor {
               }
             }
           }
-          else if (firstcode == 2) {
-            // search for a load instruction
+          else {  // firstCode == 2
+            // searching for a load instruction
             if (exprent.type == Exprent.EXPRENT_ASSIGNMENT) {
-              AssignmentExprent assexpr = (AssignmentExprent)exprent;
-              if (assexpr.getRight().type == Exprent.EXPRENT_VAR &&
-                  new VarVersionPair((VarExprent)assexpr.getRight()).equals(varpaar)) {
-
+              AssignmentExprent assignment = (AssignmentExprent)exprent;
+              if (assignment.getRight().type == Exprent.EXPRENT_VAR &&
+                  new VarVersionPair((VarExprent)assignment.getRight()).equals(pair)) {
                 Exprent next = null;
                 if (i == node.exprents.size() - 1) {
                   if (node.successors.size() == 1) {
@@ -228,9 +223,9 @@ public class FinallyProcessor {
 
                 boolean found = false;
                 if (next != null && next.type == Exprent.EXPRENT_EXIT) {
-                  ExitExprent exexpr = (ExitExprent)next;
-                  if (exexpr.getExitType() == ExitExprent.EXIT_THROW && exexpr.getValue().type == Exprent.EXPRENT_VAR
-                      && assexpr.getLeft().equals(exexpr.getValue())) {
+                  ExitExprent exit = (ExitExprent)next;
+                  if (exit.getExitType() == ExitExprent.EXIT_THROW && exit.getValue().type == Exprent.EXPRENT_VAR &&
+                      assignment.getLeft().equals(exit.getValue())) {
                     found = true;
                   }
                 }
@@ -266,30 +261,24 @@ public class FinallyProcessor {
       stack.addAll(node.successors);
     }
 
-    // empty finally block?
+    // an empty `finally` block?
     if (fstat.getHandler().type == StatementType.BASIC_BLOCK) {
-
-      boolean isEmpty = false;
       boolean isFirstLast = mapLast.containsKey(firstBasicBlock);
       InstructionSequence seq = firstBasicBlock.getSeq();
 
-      switch (firstcode) {
-        case 0:
-          isEmpty = isFirstLast && seq.length() == 1;
-          break;
-        case 1:
-          isEmpty = seq.length() == 1;
-          break;
-        case 2:
-          isEmpty = isFirstLast ? seq.length() == 3 : seq.length() == 1;
-      }
+      boolean isEmpty = switch (firstCode) {
+        case 0 -> isFirstLast && seq.length() == 1;
+        case 1 -> seq.length() == 1;
+        case 2 -> isFirstLast ? seq.length() == 3 : seq.length() == 1;
+        default -> false;
+      };
 
       if (isEmpty) {
-        firstcode = 3;
+        firstCode = 3;
       }
     }
 
-    return new Record(firstcode, mapLast);
+    return new Record(firstCode, mapLast);
   }
 
   private static void insertSemaphore(ControlFlowGraph graph,
@@ -298,34 +287,32 @@ public class FinallyProcessor {
                                       BasicBlock handler,
                                       int var,
                                       Record information,
-                                      int bytecode_version) {
+                                      int bytecodeVersion) {
     Set<BasicBlock> setCopy = new HashSet<>(setTry);
 
-    int finallytype = information.firstCode;
+    int finallyType = information.firstCode;
     Map<BasicBlock, Boolean> mapLast = information.mapLast;
 
     // first and last statements
-    removeExceptionInstructionsEx(handler, 1, finallytype);
+    removeExceptionInstructionsEx(handler, 1, finallyType);
     for (Entry<BasicBlock, Boolean> entry : mapLast.entrySet()) {
       BasicBlock last = entry.getKey();
 
       if (entry.getValue()) {
-        removeExceptionInstructionsEx(last, 2, finallytype);
+        removeExceptionInstructionsEx(last, 2, finallyType);
         graph.getFinallyExits().add(last);
       }
     }
 
     // disable semaphore at statement exit points
     for (BasicBlock block : setTry) {
-      List<BasicBlock> lstSucc = block.getSuccessors();
-
-      for (BasicBlock dest : lstSucc) {
+      for (BasicBlock dest : block.getSuccessors()) {
         // break out
         if (dest != graph.getLast() && !setCopy.contains(dest)) {
           // disable semaphore
           SimpleInstructionSequence seq = new SimpleInstructionSequence();
-          seq.addInstruction(Instruction.create(CodeConstants.opc_bipush, false, CodeConstants.GROUP_GENERAL, bytecode_version, new int[]{0}), -1);
-          seq.addInstruction(Instruction.create(CodeConstants.opc_istore, false, CodeConstants.GROUP_GENERAL, bytecode_version, new int[]{var}), -1);
+          seq.addInstruction(Instruction.create(CodeConstants.opc_bipush, false, CodeConstants.GROUP_GENERAL, bytecodeVersion, new int[]{0}), -1);
+          seq.addInstruction(Instruction.create(CodeConstants.opc_istore, false, CodeConstants.GROUP_GENERAL, bytecodeVersion, new int[]{var}), -1);
 
           // build a separate block
           BasicBlock newBlock = new BasicBlock(++graph.last_id, seq);
@@ -338,35 +325,17 @@ public class FinallyProcessor {
 
           // exception ranges
           // FIXME: special case synchronized
-
-          // copy exception edges and extend protected ranges
-          for (int j = 0; j < block.getSuccessorExceptions().size(); j++) {
-            BasicBlock hd = block.getSuccessorExceptions().get(j);
-            newBlock.addSuccessorException(hd);
-
-            ExceptionRangeCFG range = graph.getExceptionRange(hd, block);
-            range.getProtectedRange().add(newBlock);
-          }
+          copyExceptionEdges(graph, block, newBlock);
         }
       }
     }
 
     // enable semaphore at the statement entrance
-    SimpleInstructionSequence seq = new SimpleInstructionSequence();
-    seq.addInstruction(Instruction.create(CodeConstants.opc_bipush, false, CodeConstants.GROUP_GENERAL, bytecode_version, new int[]{1}), -1);
-    seq.addInstruction(Instruction.create(CodeConstants.opc_istore, false, CodeConstants.GROUP_GENERAL, bytecode_version, new int[]{var}), -1);
-
-    BasicBlock newHead = new BasicBlock(++graph.last_id, seq);
-
+    BasicBlock newHead = createHeadBlock(graph, 1, var, bytecodeVersion);
     insertBlockBefore(graph, head, newHead);
 
     // initialize semaphore with false
-    seq = new SimpleInstructionSequence();
-    seq.addInstruction(Instruction.create(CodeConstants.opc_bipush, false, CodeConstants.GROUP_GENERAL, bytecode_version, new int[]{0}), -1);
-    seq.addInstruction(Instruction.create(CodeConstants.opc_istore, false, CodeConstants.GROUP_GENERAL, bytecode_version, new int[]{var}), -1);
-
-    BasicBlock newHeadInit = new BasicBlock(++graph.last_id, seq);
-
+    BasicBlock newHeadInit = createHeadBlock(graph, 0, var, bytecodeVersion);
     insertBlockBefore(graph, newHead, newHeadInit);
 
     setCopy.add(newHead);
@@ -382,35 +351,43 @@ public class FinallyProcessor {
     }
   }
 
-  private static void insertBlockBefore(ControlFlowGraph graph, BasicBlock oldblock, BasicBlock newblock) {
-    List<BasicBlock> lstTemp = new ArrayList<>();
-    lstTemp.addAll(oldblock.getPredecessors());
-    lstTemp.addAll(oldblock.getPredecessorExceptions());
+  @NotNull
+  private static BasicBlock createHeadBlock(ControlFlowGraph graph, int value, int var, int bytecodeVersion) {
+    SimpleInstructionSequence seq = new SimpleInstructionSequence();
+    seq.addInstruction(Instruction.create(CodeConstants.opc_bipush, false, CodeConstants.GROUP_GENERAL, bytecodeVersion, new int[]{value}), -1);
+    seq.addInstruction(Instruction.create(CodeConstants.opc_istore, false, CodeConstants.GROUP_GENERAL, bytecodeVersion, new int[]{var}), -1);
+    return new BasicBlock(++graph.last_id, seq);
+  }
+
+  private static void insertBlockBefore(ControlFlowGraph graph, BasicBlock oldBlock, BasicBlock newBlock) {
+    List<BasicBlock> blocks = new ArrayList<>();
+    blocks.addAll(oldBlock.getPredecessors());
+    blocks.addAll(oldBlock.getPredecessorExceptions());
 
     // replace predecessors
-    for (BasicBlock pred : lstTemp) {
-      pred.replaceSuccessor(oldblock, newblock);
+    for (BasicBlock predecessor : blocks) {
+      predecessor.replaceSuccessor(oldBlock, newBlock);
     }
 
     // copy exception edges and extend protected ranges
-    for (BasicBlock hd : oldblock.getSuccessorExceptions()) {
-      newblock.addSuccessorException(hd);
+    for (BasicBlock hd : oldBlock.getSuccessorExceptions()) {
+      newBlock.addSuccessorException(hd);
 
-      ExceptionRangeCFG range = graph.getExceptionRange(hd, oldblock);
-      range.getProtectedRange().add(newblock);
+      ExceptionRangeCFG range = graph.getExceptionRange(hd, oldBlock);
+      range.getProtectedRange().add(newBlock);
     }
 
     // replace handler
     for (ExceptionRangeCFG range : graph.getExceptions()) {
-      if (range.getHandler() == oldblock) {
-        range.setHandler(newblock);
+      if (range.getHandler() == oldBlock) {
+        range.setHandler(newBlock);
       }
     }
 
-    newblock.addSuccessor(oldblock);
-    graph.getBlocks().addWithKey(newblock, newblock.id);
-    if (graph.getFirst() == oldblock) {
-      graph.setFirst(newblock);
+    newBlock.addSuccessor(oldBlock);
+    graph.getBlocks().addWithKey(newBlock, newBlock.id);
+    if (graph.getFirst() == oldBlock) {
+      graph.setFirst(newBlock);
     }
   }
 
@@ -421,7 +398,6 @@ public class FinallyProcessor {
     int index = 0;
     do {
       Statement st = lst.get(index);
-
       if (st.type == StatementType.BASIC_BLOCK) {
         index++;
       }
@@ -433,11 +409,9 @@ public class FinallyProcessor {
     while (index < lst.size());
 
     Set<BasicBlock> res = new HashSet<>();
-
     for (Statement st : lst) {
       res.add(((BasicBlockStatement)st).getBlock());
     }
-
     return res;
   }
 
@@ -445,15 +419,15 @@ public class FinallyProcessor {
     Set<BasicBlock> tryBlocks = getAllBasicBlocks(fstat.getFirst());
     Set<BasicBlock> catchBlocks = getAllBasicBlocks(fstat.getHandler());
 
-    int finallytype = information.firstCode;
+    int finallyType = information.firstCode;
     Map<BasicBlock, Boolean> mapLast = information.mapLast;
 
     BasicBlock first = fstat.getHandler().getBasichead().getBlock();
     boolean skippedFirst = false;
 
-    if (finallytype == 3) {
+    if (finallyType == 3) {
       // empty finally
-      removeExceptionInstructionsEx(first, 3, finallytype);
+      removeExceptionInstructionsEx(first, 3, finallyType);
 
       if (mapLast.containsKey(first)) {
         graph.getFinallyExits().add(first);
@@ -462,10 +436,10 @@ public class FinallyProcessor {
       return true;
     }
     else {
-      if (first.getSeq().length() == 1 && finallytype > 0) {
-        BasicBlock firstsuc = first.getSuccessors().get(0);
-        if (catchBlocks.contains(firstsuc)) {
-          first = firstsuc;
+      if (first.getSeq().length() == 1 && finallyType > 0) {
+        BasicBlock firstSuccessor = first.getSuccessors().get(0);
+        if (catchBlocks.contains(firstSuccessor)) {
+          first = firstSuccessor;
           skippedFirst = true;
         }
       }
@@ -476,47 +450,35 @@ public class FinallyProcessor {
     for (BasicBlock block : tryBlocks) {
       startBlocks.addAll(block.getSuccessors());
     }
-    // throw in the try body will point directly to the dummy exit
-    // so remove dummy exit
+    // `throw` in the `try` body will point directly to the dummy exit, so remove it
     startBlocks.remove(graph.getLast());
     startBlocks.removeAll(tryBlocks);
 
-    List<Area> lstAreas = new ArrayList<>();
-
+    List<Area> areas = new ArrayList<>();
     for (BasicBlock start : startBlocks) {
-
-      Area arr = compareSubgraphsEx(graph, start, catchBlocks, first, finallytype, mapLast, skippedFirst);
+      Area arr = compareSubGraphsEx(graph, start, catchBlocks, first, finallyType, mapLast, skippedFirst);
       if (arr == null) {
         return false;
       }
-
-      lstAreas.add(arr);
+      areas.add(arr);
     }
 
-    //		try {
-    //			DotExporter.toDotFile(graph, new File("c:\\Temp\\fern5.dot"), true);
-    //		} catch(Exception ex){ex.printStackTrace();}
-
     // delete areas
-    for (Area area : lstAreas) {
+    for (Area area : areas) {
       deleteArea(graph, area);
     }
 
-    //		try {
-    //			DotExporter.toDotFile(graph, new File("c:\\Temp\\fern5.dot"), true);
-    //		} catch(Exception ex){ex.printStackTrace();}
-
-    // INFO: empty basic blocks may remain in the graph!
+    // INFO: Empty basic blocks may remain in the graph!
     for (Entry<BasicBlock, Boolean> entry : mapLast.entrySet()) {
       BasicBlock last = entry.getKey();
 
       if (entry.getValue()) {
-        removeExceptionInstructionsEx(last, 2, finallytype);
+        removeExceptionInstructionsEx(last, 2, finallyType);
         graph.getFinallyExits().add(last);
       }
     }
 
-    removeExceptionInstructionsEx(fstat.getHandler().getBasichead().getBlock(), 1, finallytype);
+    removeExceptionInstructionsEx(fstat.getHandler().getBasichead().getBlock(), 1, finallyType);
 
     return true;
   }
@@ -533,11 +495,11 @@ public class FinallyProcessor {
     }
   }
 
-  private Area compareSubgraphsEx(ControlFlowGraph graph,
+  private Area compareSubGraphsEx(ControlFlowGraph graph,
                                   BasicBlock startSample,
                                   Set<BasicBlock> catchBlocks,
                                   BasicBlock startCatch,
-                                  int finallytype,
+                                  int finallyType,
                                   Map<BasicBlock, Boolean> mapLast,
                                   boolean skippedFirst) {
     class BlockStackEntry {
@@ -563,7 +525,6 @@ public class FinallyProcessor {
     stack.add(new BlockStackEntry(startCatch, startSample, new ArrayList<>()));
 
     while (!stack.isEmpty()) {
-
       BlockStackEntry entry = stack.remove(0);
       BasicBlock blockCatch = entry.blockCatch;
       BasicBlock blockSample = entry.blockSample;
@@ -572,8 +533,8 @@ public class FinallyProcessor {
       boolean isLastBlock = mapLast.containsKey(blockCatch);
       boolean isTrueLastBlock = isLastBlock && mapLast.get(blockCatch);
 
-      if (!compareBasicBlocksEx(graph, blockCatch, blockSample, (isFirstBlock ? 1 : 0) | (isTrueLastBlock ? 2 : 0), finallytype,
-                                entry.lstStoreVars)) {
+      int compareType = (isFirstBlock ? 1 : 0) | (isTrueLastBlock ? 2 : 0);
+      if (!compareBasicBlocksEx(graph, blockCatch, blockSample, compareType, finallyType, entry.lstStoreVars)) {
         return null;
       }
 
@@ -607,11 +568,8 @@ public class FinallyProcessor {
             String excSample = graph.getExceptionRange(sucSample, blockSample).getUniqueExceptionsString();
 
             // FIXME: compare handlers if possible
-            boolean equalexc = excCatch == null ? excSample == null : excCatch.equals(excSample);
-
-            if (equalexc) {
+            if (Objects.equals(excCatch, excSample)) {
               if (catchBlocks.contains(sucCatch) && !setSample.contains(sucSample)) {
-
                 List<int[]> lst = entry.lstStoreVars;
 
                 if (sucCatch.getSeq().length() > 0 && sucSample.getSeq().length() > 0) {
@@ -639,16 +597,16 @@ public class FinallyProcessor {
       }
 
       if (isLastBlock) {
-        Set<BasicBlock> setSuccs = new HashSet<>(blockSample.getSuccessors());
-        setSuccs.removeAll(setSample);
+        Set<BasicBlock> successors = new HashSet<>(blockSample.getSuccessors());
+        successors.removeAll(setSample);
 
-        for (BlockStackEntry stackent : stack) {
-          setSuccs.remove(stackent.blockSample);
+        for (BlockStackEntry stackEntry : stack) {
+          successors.remove(stackEntry.blockSample);
         }
 
-        for (BasicBlock succ : setSuccs) {
-          if (graph.getLast() != succ) { // FIXME: why?
-            mapNext.put(blockSample.id + "#" + succ.id, new BasicBlock[]{blockSample, succ, isTrueLastBlock ? succ : null});
+        for (BasicBlock successor : successors) {
+          if (graph.getLast() != successor) { // FIXME: why?
+            mapNext.put(blockSample.id + "#" + successor.id, new BasicBlock[]{blockSample, successor, isTrueLastBlock ? successor : null});
           }
         }
       }
@@ -658,13 +616,11 @@ public class FinallyProcessor {
   }
 
   private static BasicBlock getUniqueNext(ControlFlowGraph graph, Set<BasicBlock[]> setNext) {
-    // precondition: there is at most one true exit path in a finally statement
-
+    // precondition: there is at most one true exit path in the `finally` statement
     BasicBlock next = null;
     boolean multiple = false;
 
     for (BasicBlock[] arr : setNext) {
-
       if (arr[2] != null) {
         next = arr[1];
         multiple = false;
@@ -717,13 +673,6 @@ public class FinallyProcessor {
           }
         }
       }
-
-      //			try {
-      //				DotExporter.toDotFile(graph, new File("c:\\Temp\\fern5.dot"), true);
-      //			} catch(IOException ex) {
-      //				ex.printStackTrace();
-      //			}
-
       for (BasicBlock[] arr : setNext) {
         if (arr[1] != next) {
           // FIXME: exception edge possible?
@@ -742,7 +691,7 @@ public class FinallyProcessor {
                                        BasicBlock pattern,
                                        BasicBlock sample,
                                        int type,
-                                       int finallytype,
+                                       int finallyType,
                                        List<int[]> lstStoreVars) {
     InstructionSequence seqPattern = pattern.getSeq();
     InstructionSequence seqSample = sample.getSeq();
@@ -751,17 +700,17 @@ public class FinallyProcessor {
       seqPattern = seqPattern.clone();
 
       if ((type & 1) > 0) { // first
-        if (finallytype > 0) {
+        if (finallyType > 0) {
           seqPattern.removeInstruction(0);
         }
       }
 
       if ((type & 2) > 0) { // last
-        if (finallytype == 0 || finallytype == 2) {
+        if (finallyType == 0 || finallyType == 2) {
           seqPattern.removeLast();
         }
 
-        if (finallytype == 2) {
+        if (finallyType == 2) {
           seqPattern.removeLast();
         }
       }
@@ -811,17 +760,20 @@ public class FinallyProcessor {
         setFinallyExits.add(newBlock);
       }
 
-      // copy exception edges and extend protected ranges
-      for (int j = 0; j < sample.getSuccessorExceptions().size(); j++) {
-        BasicBlock hd = sample.getSuccessorExceptions().get(j);
-        newBlock.addSuccessorException(hd);
-
-        ExceptionRangeCFG range = graph.getExceptionRange(hd, sample);
-        range.getProtectedRange().add(newBlock);
-      }
+      copyExceptionEdges(graph, sample, newBlock);
     }
 
     return true;
+  }
+
+  // copy exception edges and extend protected ranges
+  public static void copyExceptionEdges(ControlFlowGraph graph, BasicBlock sample, BasicBlock newBlock) {
+    for (int i = 0; i < sample.getSuccessorExceptions().size(); i++) {
+      BasicBlock hd = sample.getSuccessorExceptions().get(i);
+      newBlock.addSuccessorException(hd);
+      ExceptionRangeCFG range = graph.getExceptionRange(hd, sample);
+      range.getProtectedRange().add(newBlock);
+    }
   }
 
   public boolean equalInstructions(Instruction first, Instruction second, List<int[]> lstStoreVars) {
@@ -864,34 +816,31 @@ public class FinallyProcessor {
       next = graph.getLast();
     }
 
-    // collect common exception ranges of predecessors and successors
+    // collecting common exception ranges of predecessors and successors
     Set<BasicBlock> setCommonExceptionHandlers = new HashSet<>(next.getSuccessorExceptions());
-    for (BasicBlock pred : start.getPredecessors()) {
-      setCommonExceptionHandlers.retainAll(pred.getSuccessorExceptions());
+    for (BasicBlock predecessor : start.getPredecessors()) {
+      setCommonExceptionHandlers.retainAll(predecessor.getSuccessorExceptions());
     }
 
-    boolean is_outside_range = false;
+    boolean isOutsideRange = false;
 
     Set<BasicBlock> setPredecessors = new HashSet<>(start.getPredecessors());
 
     // replace start with next
-    for (BasicBlock pred : setPredecessors) {
-      pred.replaceSuccessor(start, next);
+    for (BasicBlock predecessor : setPredecessors) {
+      predecessor.replaceSuccessor(start, next);
     }
 
     Set<BasicBlock> setBlocks = area.sample;
 
     Set<ExceptionRangeCFG> setCommonRemovedExceptionRanges = null;
 
-    // remove all the blocks inbetween
+    // remove all the blocks in between
     for (BasicBlock block : setBlocks) {
-
-      // artificial basic blocks (those resulted from splitting)
-      // can belong to more than one area
+      // artificial basic blocks (those resulting from splitting) may belong to more than one area
       if (graph.getBlocks().containsKey(block.id)) {
-
-        if (!block.getSuccessorExceptions().containsAll(setCommonExceptionHandlers)) {
-          is_outside_range = true;
+        if (!new HashSet<>(block.getSuccessorExceptions()).containsAll(setCommonExceptionHandlers)) {
+          isOutsideRange = true;
         }
 
         Set<ExceptionRangeCFG> setRemovedExceptionRanges = new HashSet<>();
@@ -906,17 +855,17 @@ public class FinallyProcessor {
           setCommonRemovedExceptionRanges.retainAll(setRemovedExceptionRanges);
         }
 
-        // shift extern edges on splitted blocks
+        // shift extern edges on split blocks
         if (block.getSeq().isEmpty() && block.getSuccessors().size() == 1) {
-          BasicBlock succs = block.getSuccessors().get(0);
-          for (BasicBlock pred : new ArrayList<>(block.getPredecessors())) {
-            if (!setBlocks.contains(pred)) {
-              pred.replaceSuccessor(block, succs);
+          BasicBlock successor = block.getSuccessors().get(0);
+          for (BasicBlock predecessor : new ArrayList<>(block.getPredecessors())) {
+            if (!setBlocks.contains(predecessor)) {
+              predecessor.replaceSuccessor(block, successor);
             }
           }
 
           if (graph.getFirst() == block) {
-            graph.setFirst(succs);
+            graph.setFirst(successor);
           }
         }
 
@@ -924,47 +873,46 @@ public class FinallyProcessor {
       }
     }
 
-    if (is_outside_range) {
+    if (isOutsideRange) {
       // new empty block
-      BasicBlock emptyblock = new BasicBlock(++graph.last_id);
-
-      graph.getBlocks().addWithKey(emptyblock, emptyblock.id);
+      BasicBlock emptyBlock = new BasicBlock(++graph.last_id);
+      graph.getBlocks().addWithKey(emptyBlock, emptyBlock.id);
 
       // add to ranges if necessary
       for (ExceptionRangeCFG range : setCommonRemovedExceptionRanges) {
-        emptyblock.addSuccessorException(range.getHandler());
-        range.getProtectedRange().add(emptyblock);
+        emptyBlock.addSuccessorException(range.getHandler());
+        range.getProtectedRange().add(emptyBlock);
       }
 
       // insert between predecessors and next
-      emptyblock.addSuccessor(next);
-      for (BasicBlock pred : setPredecessors) {
-        pred.replaceSuccessor(next, emptyblock);
+      emptyBlock.addSuccessor(next);
+      for (BasicBlock predecessor : setPredecessors) {
+        predecessor.replaceSuccessor(next, emptyBlock);
       }
     }
   }
 
-  private static void removeExceptionInstructionsEx(BasicBlock block, int blocktype, int finallytype) {
+  private static void removeExceptionInstructionsEx(BasicBlock block, int blockType, int finallyType) {
     InstructionSequence seq = block.getSeq();
 
-    if (finallytype == 3) { // empty finally handler
+    if (finallyType == 3) { // empty finally handler
       for (int i = seq.length() - 1; i >= 0; i--) {
         seq.removeInstruction(i);
       }
     }
     else {
-      if ((blocktype & 1) > 0) { // first
-        if (finallytype == 2 || finallytype == 1) { // astore or pop
+      if ((blockType & 1) > 0) { // first
+        if (finallyType == 2 || finallyType == 1) { // `AStore` or `Pop`
           seq.removeInstruction(0);
         }
       }
 
-      if ((blocktype & 2) > 0) { // last
-        if (finallytype == 2 || finallytype == 0) {
+      if ((blockType & 2) > 0) { // last
+        if (finallyType == 2 || finallyType == 0) {
           seq.removeLast();
         }
 
-        if (finallytype == 2) { // astore
+        if (finallyType == 2) { // `AStore`
           seq.removeLast();
         }
       }
