@@ -72,23 +72,16 @@ public class InvocationExprent extends Exprent {
     className = cn.classname;
     this.bootstrapArguments = bootstrapArguments;
     switch (opcode) {
-      case CodeConstants.opc_invokestatic:
-        invocationType = INVOKE_STATIC;
-        break;
-      case CodeConstants.opc_invokespecial:
-        invocationType = INVOKE_SPECIAL;
-        break;
-      case CodeConstants.opc_invokevirtual:
-        invocationType = INVOKE_VIRTUAL;
-        break;
-      case CodeConstants.opc_invokeinterface:
-        invocationType = INVOKE_INTERFACE;
-        break;
-      case CodeConstants.opc_invokedynamic:
+      case CodeConstants.opc_invokestatic -> invocationType = INVOKE_STATIC;
+      case CodeConstants.opc_invokespecial -> invocationType = INVOKE_SPECIAL;
+      case CodeConstants.opc_invokevirtual -> invocationType = INVOKE_VIRTUAL;
+      case CodeConstants.opc_invokeinterface -> invocationType = INVOKE_INTERFACE;
+      case CodeConstants.opc_invokedynamic -> {
         invocationType = INVOKE_DYNAMIC;
 
         className = "java/lang/Class"; // dummy class name
         invokeDynamicClassSuffix = "##Lambda_" + cn.index1 + "_" + cn.index2;
+      }
     }
 
     if (CodeConstants.INIT_NAME.equals(name)) {
@@ -428,27 +421,17 @@ public class InvocationExprent extends Exprent {
 
   // TODO: move to CodeConstants ???
   private static String getClassNameForPrimitiveType(int type) {
-    switch (type) {
-      case CodeConstants.TYPE_BOOLEAN:
-        return "java/lang/Boolean";
-      case CodeConstants.TYPE_BYTE:
-      case CodeConstants.TYPE_BYTECHAR:
-        return "java/lang/Byte";
-      case CodeConstants.TYPE_CHAR:
-        return "java/lang/Character";
-      case CodeConstants.TYPE_SHORT:
-      case CodeConstants.TYPE_SHORTCHAR:
-        return "java/lang/Short";
-      case CodeConstants.TYPE_INT:
-        return "java/lang/Integer";
-      case CodeConstants.TYPE_LONG:
-        return "java/lang/Long";
-      case CodeConstants.TYPE_FLOAT:
-        return "java/lang/Float";
-      case CodeConstants.TYPE_DOUBLE:
-        return "java/lang/Double";
-    }
-    return null;
+    return switch (type) {
+      case CodeConstants.TYPE_BOOLEAN -> "java/lang/Boolean";
+      case CodeConstants.TYPE_BYTE, CodeConstants.TYPE_BYTECHAR -> "java/lang/Byte";
+      case CodeConstants.TYPE_CHAR -> "java/lang/Character";
+      case CodeConstants.TYPE_SHORT, CodeConstants.TYPE_SHORTCHAR -> "java/lang/Short";
+      case CodeConstants.TYPE_INT -> "java/lang/Integer";
+      case CodeConstants.TYPE_LONG -> "java/lang/Long";
+      case CodeConstants.TYPE_FLOAT -> "java/lang/Float";
+      case CodeConstants.TYPE_DOUBLE -> "java/lang/Double";
+      default -> null;
+    };
   }
 
   private static final Map<String, String> UNBOXING_METHODS = Map.of(
@@ -537,13 +520,13 @@ public class InvocationExprent extends Exprent {
     if (!(o instanceof InvocationExprent)) return false;
 
     InvocationExprent it = (InvocationExprent)o;
-    return Objects.equals(name, it.getName()) &&
-           Objects.equals(className, it.getClassName()) &&
-           isStatic == it.isStatic() &&
-           Objects.equals(instance, it.getInstance()) &&
-           Objects.equals(descriptor, it.getDescriptor()) &&
-           funcType == it.getFuncType() &&
-           InterpreterUtil.equalLists(parameters, it.getParameters());
+    return Objects.equals(name, it.name) &&
+           Objects.equals(className, it.className) &&
+           isStatic == it.isStatic &&
+           Objects.equals(instance, it.instance) &&
+           Objects.equals(descriptor, it.descriptor) &&
+           funcType == it.funcType &&
+           Objects.equals(parameters, it.parameters);
   }
 
   public List<Exprent> getParameters() {
