@@ -66,30 +66,30 @@ public final class DoStatement extends Statement {
       tracer.incrementCurrentSourceLine();
     }
     switch (loopType) {
-      case DO:
+      case DO -> {
         buf.appendIndent(indent).append("while(true) {").appendLineSeparator();
         tracer.incrementCurrentSourceLine();
         buf.append(ExprProcessor.jmpWrapper(first, indent + 1, false, tracer));
         buf.appendIndent(indent).append("}").appendLineSeparator();
         tracer.incrementCurrentSourceLine();
-        break;
-      case DO_WHILE:
+      }
+      case DO_WHILE -> {
         buf.appendIndent(indent).append("do {").appendLineSeparator();
         tracer.incrementCurrentSourceLine();
         buf.append(ExprProcessor.jmpWrapper(first, indent + 1, false, tracer));
         buf.appendIndent(indent).append("} while(").append(
           Objects.requireNonNull(conditionExprent.get(0)).toJava(indent, tracer)).append(");").appendLineSeparator();
         tracer.incrementCurrentSourceLine();
-        break;
-      case WHILE:
+      }
+      case WHILE -> {
         buf.appendIndent(indent).append("while(").append(
           Objects.requireNonNull(conditionExprent.get(0)).toJava(indent, tracer)).append(") {").appendLineSeparator();
         tracer.incrementCurrentSourceLine();
         buf.append(ExprProcessor.jmpWrapper(first, indent + 1, false, tracer));
         buf.appendIndent(indent).append("}").appendLineSeparator();
         tracer.incrementCurrentSourceLine();
-        break;
-      case FOR:
+      }
+      case FOR -> {
         buf.appendIndent(indent).append("for(");
         Exprent firstInitExprent = initExprent.get(0);
         if (firstInitExprent != null) {
@@ -97,12 +97,14 @@ public final class DoStatement extends Statement {
         }
         Exprent firstIncExprent = Objects.requireNonNull(incExprent.get(0));
         buf.append("; ")
-          .append(Objects.requireNonNull(conditionExprent.get(0)).toJava(indent, tracer)).append("; ").append(firstIncExprent.toJava(indent, tracer)).append(") {")
+          .append(Objects.requireNonNull(conditionExprent.get(0)).toJava(indent, tracer)).append("; ")
+          .append(firstIncExprent.toJava(indent, tracer)).append(") {")
           .appendLineSeparator();
         tracer.incrementCurrentSourceLine();
         buf.append(ExprProcessor.jmpWrapper(first, indent + 1, false, tracer));
         buf.appendIndent(indent).append("}").appendLineSeparator();
         tracer.incrementCurrentSourceLine();
+      }
     }
     return buf;
   }
@@ -120,11 +122,8 @@ public final class DoStatement extends Statement {
     }
     lst.add(first);
     switch (loopType) {
-      case DO_WHILE:
-        lst.add(getConditionExprent());
-        break;
-      case FOR:
-        lst.add(getIncExprent());
+      case DO_WHILE -> lst.add(getConditionExprent());
+      case FOR -> lst.add(getIncExprent());
     }
     return lst;
   }
