@@ -17,16 +17,10 @@ public final class Statements {
       return null;
     }
 
-    switch (stat.type) {
-      case SEQUENCE:
-      case IF:
-      case ROOT:
-      case SWITCH:
-      case SYNCHRONIZED:
-        return findFirstData(stat.getFirst());
-      default:
-        return null;
-    }
+    return switch (stat.type) {
+      case SEQUENCE, IF, ROOT, SWITCH, SYNCHRONIZED -> findFirstData(stat.getFirst());
+      default -> null;
+    };
   }
 
   public static boolean isInvocationInitConstructor(InvocationExprent inv, MethodWrapper method, ClassWrapper wrapper, boolean withThis) {
