@@ -179,60 +179,28 @@ public class VarVersionsProcessor {
                              VarType.getCommonMinType(firstMaxType, secondMaxType);
 
               if (firstType.getTypeFamily() == CodeConstants.TYPE_FAMILY_INTEGER && secondType.getTypeFamily() == CodeConstants.TYPE_FAMILY_INTEGER) {
-                switch (secondType.getType()) {
-                  case CodeConstants.TYPE_INT:
-                    type = VarType.VARTYPE_INT;
-                    break;
-                  case CodeConstants.TYPE_SHORT:
-                    type = firstType.getType() == CodeConstants.TYPE_INT ? null : VarType.VARTYPE_SHORT;
-                    break;
-                  case CodeConstants.TYPE_CHAR:
-                    switch (firstType.getType()) {
-                      case CodeConstants.TYPE_INT:
-                      case CodeConstants.TYPE_SHORT:
-                        type = null;
-                        break;
-                      default:
-                        type = VarType.VARTYPE_CHAR;
-                    }
-                    break;
-                  case CodeConstants.TYPE_SHORTCHAR:
-                    switch (firstType.getType()) {
-                      case CodeConstants.TYPE_INT:
-                      case CodeConstants.TYPE_SHORT:
-                      case CodeConstants.TYPE_CHAR:
-                        type = null;
-                        break;
-                      default:
-                        type = VarType.VARTYPE_SHORTCHAR;
-                    }
-                    break;
-                  case CodeConstants.TYPE_BYTECHAR:
-                    switch (firstType.getType()) {
-                      case CodeConstants.TYPE_INT:
-                      case CodeConstants.TYPE_SHORT:
-                      case CodeConstants.TYPE_CHAR:
-                      case CodeConstants.TYPE_SHORTCHAR:
-                        type = null;
-                        break;
-                      default:
-                        type = VarType.VARTYPE_BYTECHAR;
-                    }
-                    break;
-                  case CodeConstants.TYPE_BYTE:
-                    switch (firstType.getType()) {
-                      case CodeConstants.TYPE_INT:
-                      case CodeConstants.TYPE_SHORT:
-                      case CodeConstants.TYPE_CHAR:
-                      case CodeConstants.TYPE_SHORTCHAR:
-                      case CodeConstants.TYPE_BYTECHAR:
-                        type = null;
-                        break;
-                      default:
-                        type = VarType.VARTYPE_BYTE;
-                    }
-                    break;
-                }
+                type = switch (secondType.getType()) {
+                  case CodeConstants.TYPE_INT -> VarType.VARTYPE_INT;
+                  case CodeConstants.TYPE_SHORT -> firstType.getType() == CodeConstants.TYPE_INT ? null : VarType.VARTYPE_SHORT;
+                  case CodeConstants.TYPE_CHAR -> switch (firstType.getType()) {
+                    case CodeConstants.TYPE_INT, CodeConstants.TYPE_SHORT -> null;
+                    default -> VarType.VARTYPE_CHAR;
+                  };
+                  case CodeConstants.TYPE_SHORTCHAR -> switch (firstType.getType()) {
+                    case CodeConstants.TYPE_INT, CodeConstants.TYPE_SHORT, CodeConstants.TYPE_CHAR -> null;
+                    default -> VarType.VARTYPE_SHORTCHAR;
+                  };
+                  case CodeConstants.TYPE_BYTECHAR -> switch (firstType.getType()) {
+                    case CodeConstants.TYPE_INT, CodeConstants.TYPE_SHORT, CodeConstants.TYPE_CHAR, CodeConstants.TYPE_SHORTCHAR -> null;
+                    default -> VarType.VARTYPE_BYTECHAR;
+                  };
+                  case CodeConstants.TYPE_BYTE -> switch (firstType.getType()) {
+                    case CodeConstants.TYPE_INT, CodeConstants.TYPE_SHORT, CodeConstants.TYPE_CHAR, CodeConstants.TYPE_SHORTCHAR, CodeConstants.TYPE_BYTECHAR ->
+                      null;
+                    default -> VarType.VARTYPE_BYTE;
+                  };
+                  default -> type;
+                };
                 if (type == null) {
                   continue;
                 }
