@@ -84,7 +84,7 @@ public class ContextUnit {
 
   public void save() {
     switch (type) {
-      case TYPE_FOLDER:
+      case TYPE_FOLDER -> {
         // create folder
         resultSaver.saveFolder(filename);
 
@@ -111,11 +111,8 @@ public class ContextUnit {
             }
           }
         }
-
-        break;
-
-      case TYPE_JAR:
-      case TYPE_ZIP:
+      }
+      case TYPE_JAR, TYPE_ZIP -> {
         // create archive file
         resultSaver.saveFolder(archivePath);
         resultSaver.createArchive(archivePath, filename, manifest);
@@ -168,6 +165,7 @@ public class ContextUnit {
         }
 
         resultSaver.closeArchive(archivePath, filename);
+      }
     }
   }
 
