@@ -116,44 +116,17 @@ public class MatchEngine {
             strValue = values[2];
           }
 
-          switch (property) {
-            case STATEMENT_TYPE:
-              value = stat_type.get(strValue);
-              break;
-            case STATEMENT_STATSIZE:
-            case STATEMENT_EXPRSIZE:
-              value = Integer.valueOf(strValue);
-              break;
-            case STATEMENT_POSITION:
-            case EXPRENT_POSITION:
-            case EXPRENT_INVOCATION_CLASS:
-            case EXPRENT_INVOCATION_SIGNATURE:
-            case EXPRENT_INVOCATION_PARAMETER:
-            case EXPRENT_VAR_INDEX:
-            case EXPRENT_FIELD_NAME:
-            case EXPRENT_CONSTVALUE:
-            case STATEMENT_RET:
-            case EXPRENT_RET:
-              value = strValue;
-              break;
-            case STATEMENT_IFTYPE:
-              value = stat_if_type.get(strValue);
-              break;
-            case EXPRENT_FUNCTYPE:
-              value = expr_func_type.get(strValue);
-              break;
-            case EXPRENT_EXITTYPE:
-              value = expr_exit_type.get(strValue);
-              break;
-            case EXPRENT_CONSTTYPE:
-              value = expr_const_type.get(strValue);
-              break;
-            case EXPRENT_TYPE:
-              value = expr_type.get(strValue);
-              break;
-            default:
-              throw new RuntimeException("Unhandled matching property");
-          }
+          value = switch (property) {
+            case STATEMENT_TYPE -> stat_type.get(strValue);
+            case STATEMENT_STATSIZE, STATEMENT_EXPRSIZE -> Integer.valueOf(strValue);
+            case STATEMENT_POSITION, EXPRENT_POSITION, EXPRENT_INVOCATION_CLASS, EXPRENT_INVOCATION_SIGNATURE, EXPRENT_INVOCATION_PARAMETER, EXPRENT_VAR_INDEX, EXPRENT_FIELD_NAME, EXPRENT_CONSTVALUE, STATEMENT_RET, EXPRENT_RET ->
+              strValue;
+            case STATEMENT_IFTYPE -> stat_if_type.get(strValue);
+            case EXPRENT_FUNCTYPE -> expr_func_type.get(strValue);
+            case EXPRENT_EXITTYPE -> expr_exit_type.get(strValue);
+            case EXPRENT_CONSTTYPE -> expr_const_type.get(strValue);
+            case EXPRENT_TYPE -> expr_type.get(strValue);
+          };
 
           matchNode.addRule(property, new RuleValue(parameter, value));
         }
