@@ -90,7 +90,7 @@ public class MethodProcessorRunnable implements Runnable {
     // TODO: move to the start, before jsr inlining
     DeadCodeHelper.connectDummyExitBlock(graph);
 
-    DeadCodeHelper.removeGotos(graph);
+    DeadCodeHelper.removeGoTos(graph);
 
     ExceptionDeobfuscator.removeCircularRanges(graph);
 
@@ -102,7 +102,7 @@ public class MethodProcessorRunnable implements Runnable {
 
     if (DecompilerContext.getOption(IFernflowerPreferences.ENSURE_SYNCHRONIZED_MONITOR)) {
       // special case: search for 'synchronized' ranges w/o monitorexit instruction (as generated by Kotlin and Scala)
-      DeadCodeHelper.extendSynchronizedRangeToMonitorexit(graph);
+      DeadCodeHelper.extendSynchronizedRangeToMonitorExit(graph);
     }
 
     if (DecompilerContext.getOption(IFernflowerPreferences.NO_EXCEPTIONS_RETURN)) {
