@@ -2,7 +2,6 @@
 package org.jetbrains.java.decompiler;
 
 import org.jetbrains.java.decompiler.main.DecompilerContext;
-import org.jetbrains.java.decompiler.main.decompiler.ConsoleDecompiler;
 import org.jetbrains.java.decompiler.main.extern.ClassFormatException;
 import org.jetbrains.java.decompiler.main.extern.IFernflowerPreferences;
 import org.junit.After;
@@ -11,12 +10,16 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.Timeout;
 
-import java.io.File;
 import java.io.IOException;
+import java.io.UncheckedIOException;
+import java.nio.file.DirectoryStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.jetbrains.java.decompiler.DecompilerTestFixture.assertFilesEqual;
 import static org.junit.Assert.assertTrue;
 
@@ -33,14 +36,14 @@ public class SingleClassesTest {
   @Before
   public void setUp() throws IOException {
     fixture = new DecompilerTestFixture();
-    fixture.setUp(IFernflowerPreferences.BYTECODE_SOURCE_MAPPING, "1",
-                  IFernflowerPreferences.DUMP_ORIGINAL_LINES, "1",
-                  IFernflowerPreferences.IGNORE_INVALID_BYTECODE, "1",
-                  IFernflowerPreferences.VERIFY_ANONYMOUS_CLASSES, "1");
+    fixture.setUp(Map.of(IFernflowerPreferences.BYTECODE_SOURCE_MAPPING, "1",
+                         IFernflowerPreferences.DUMP_ORIGINAL_LINES, "1",
+                         IFernflowerPreferences.IGNORE_INVALID_BYTECODE, "1",
+                         IFernflowerPreferences.VERIFY_ANONYMOUS_CLASSES, "1"));
   }
 
   @After
-  public void tearDown() {
+  public void tearDown() throws IOException {
     fixture.tearDown();
     fixture = null;
   }
@@ -120,6 +123,7 @@ public class SingleClassesTest {
           "pkg/SharedName2", "pkg/SharedName3", "pkg/SharedName4", "pkg/NonSharedName",
           "pkg/TestClashNameParent", "ext/TestClashNameParent","pkg/TestClashNameIface", "ext/TestClashNameIface"); }
   @Test public void testSwitchOnEnum() { doTest("pkg/TestSwitchOnEnum");}
+  @Test public void testSwitchOnEnumEclipse() { doTest("pkg/TestSwitchOnEnumEclipse"); }
   @Test public void testVarArgCalls() { doTest("pkg/TestVarArgCalls"); }
   @Test public void testLambdaParams() { doTest("pkg/TestLambdaParams"); }
   @Test public void testInterfaceMethods() { doTest("pkg/TestInterfaceMethods"); }
@@ -215,13 +219,11 @@ public class SingleClassesTest {
   @Test public void testNestedType() { doTest("pkg/NestedType"); }
   @Test public void testInheritanceChainCycle() { doTest("pkg/TestInheritanceChainCycle"); }
   @Test public void testDynamicConstantPoolEntry() { doTest("java11/TestDynamicConstantPoolEntry"); }
-
-
   @Test public void testInstanceofWithPattern() {
     doTest("patterns/TestInstanceofWithPattern");
   }
   @Test public void testInstanceofVarNotSupported() {
-    // bytecode version of this test data doesn't support patterns in instanceof, so no modifications regarding that are applied
+    // the bytecode version of this test data doesn't support patterns in `instanceof`, so no modifications regarding that are applied
     doTest("patterns/TestInstanceofPatternNotSupported");
   }
 
@@ -229,41 +231,45 @@ public class SingleClassesTest {
   public void testUnsupportedConstantPoolEntry() { doTest("java11/TestUnsupportedConstantPoolEntry"); }
 
   private void doTest(String testFile, String... companionFiles) {
-    ConsoleDecompiler decompiler = fixture.getDecompiler();
+    var decompiler = fixture.getDecompiler();
 
-    File classFile = new File(fixture.getTestDataDir(), "/classes/" + testFile + ".class");
-    assertTrue(classFile.isFile());
-    for (File file : collectClasses(classFile)) {
-      decompiler.addSource(file);
+    var classFile = fixture.getTestDataDir().resolve("classes/" + testFile + ".class");
+    assertThat(classFile).isRegularFile();
+    for (var file : collectClasses(classFile)) {
+      decompiler.addSource(file.toFile());
     }
 
     for (String companionFile : companionFiles) {
-      File companionClassFile = new File(fixture.getTestDataDir(), "/classes/" + companionFile + ".class");
-      assertTrue(companionClassFile.isFile());
-      for (File file : collectClasses(companionClassFile)) {
-        decompiler.addSource(file);
+      var companionClassFile = fixture.getTestDataDir().resolve("classes/" + companionFile + ".class");
+      assertThat(companionClassFile).isRegularFile();
+      for (var file : collectClasses(companionClassFile)) {
+        decompiler.addSource(file.toFile());
       }
     }
 
     decompiler.decompileContext();
 
-    String testName = classFile.getName().substring(0, classFile.getName().length() - 6);
-    File decompiledFile = new File(fixture.getTargetDir(), testName + ".java");
-    assertTrue(decompiledFile.isFile());
-    File referenceFile = new File(fixture.getTestDataDir(), "results/" + testName + ".dec");
-    assertTrue(referenceFile.isFile());
+    var decompiledFile = fixture.getTargetDir().resolve(classFile.getFileName().toString().replace(".class", ".java"));
+    assertThat(decompiledFile).isRegularFile();
+    assertTrue(Files.isRegularFile(decompiledFile));
+    var referenceFile = fixture.getTestDataDir().resolve("results/" + classFile.getFileName().toString().replace(".class", ".dec"));
+    assertThat(referenceFile).isRegularFile();
     assertFilesEqual(referenceFile, decompiledFile);
   }
 
-  private static List<File> collectClasses(File classFile) {
-    List<File> files = new ArrayList<>();
+  private static List<Path> collectClasses(Path classFile) {
+    var files = new ArrayList<Path>();
     files.add(classFile);
 
-    File parent = classFile.getParentFile();
+    var parent = classFile.getParent();
     if (parent != null) {
-      final String pattern = classFile.getName().replace(".class", "") + "\\$.+\\.class";
-      File[] inner = parent.listFiles((dir, name) -> name.matches(pattern));
-      if (inner != null) Collections.addAll(files, inner);
+      var glob = classFile.getFileName().toString().replace(".class", "$*.class");
+      try (DirectoryStream<Path> inner = Files.newDirectoryStream(parent, glob)) {
+        inner.forEach(files::add);
+      }
+      catch (IOException e) {
+        throw new UncheckedIOException(e);
+      }
     }
 
     return files;
